preset-env: Move all defaults to the separate module.
